moved boundary balance setup and configure to transcription base.
